% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_matrix.R
\name{choose_matrix}
\alias{choose_matrix}
\title{Choose Matrix}
\usage{
choose_matrix(sig2, rho = 0.7, p, blk_sz = 20, structure)
}
\arguments{
\item{sig2}{Variance (identical for all predictors).}

\item{rho}{Correlation used for blocks in block-diagonal structure.}

\item{p}{Number of features.}

\item{blk_sz}{Size of blocks in block-diagonal structure (default 20x20).}

\item{structure}{Covariance matrix structure (either "diag" for diagonal matrix, or "block" for block-diagonal matrix).}
}
\value{
A p x p matrix.
}
\description{
Creates a covariance matrix, either diagonal or block-diagonal, with the given parameters. The covariance matrix has identical variance across the diagonal.
}
