% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Generate Error-Prone Data}
\usage{
data_gen(
  n,
  p,
  k,
  mu = 0,
  mu_x = NULL,
  beta,
  Sig_x,
  Sig_u,
  sig2e = 0.1,
  family = "gaussian"
)
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Number of predictors.}

\item{k}{Number of replicates.}

\item{mu}{Population intercept (scalar).}

\item{mu_x}{p-dimensional mean vector for error-prone predictors.}

\item{beta}{p-dimensional vector of true coefficients (used to generate the response variable).}

\item{Sig_x}{Covariance matrix for true predictors X.}

\item{Sig_u}{Covariance matrix for measurement error U.}

\item{sig2e}{Variance of response (or, equivalently, of residual noise) in linear regression (default 0.1).}

\item{family}{Distribution of response conditional on true predictors X ("gaussian" for linear regression, or "binomial" for logistic regression).}
}
\value{
A list with true predictor matrix X, list of k error-prone replicates W, response variable y, and n, p, and k.
}
\description{
Generates a design matrix as well as its error-prone version with the specified number of replicates, along with the response varaible, based on specified parameters.
}
