% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.R
\name{rcl}
\alias{rcl}
\title{Regression-Calibrated Lasso}
\usage{
rcl(me_list, Lambda, ...)
}
\arguments{
\item{me_list}{List containing, at the very least, the number of replicates, k, the response y, and a list of k matrices of error-prone observations W -- one for each replicate.}

\item{Lambda}{Reliability matrix (or its estimate, if unknown).}

\item{...}{Inputs to the glmnet function.}
}
\value{
An object of class "glmnet".
}
\description{
Lasso for linear or logistic regression with regression calibration correction for measurement error.
}
